simplify and correct height measurement
Original function gets offsetHeight, then subtracts padding and borders from style.
But... style already contains the result of this calculation under 'height'.

Moreover - and this is the most important point - there is a problem in dialog resizing when the user has a CSS that sets box-sizing to border-box instead of the default content-box.

When box-sizing is set to border-box, style.height = value will set the OUTER height to value. Not the inner height. This causes a resized window to shrink uncontrollably as we're reading the inner height, correcting for a deltaY, and assigning the result to the outer height.

Reading style.height, correcting for deltaY, and writing to style.height makes a lot more sense than what was done before, which works or not depending on factors that we don't control.