Fix dot to comma issue in some (german) browsers
I had some issues with fading. i found out that in some of my browsers and in electron the value of opacity will have a comma instead of a dot as float separator. thats common for germany, because we write floats with comma instead of dot. maybe the local settings of the browser cause this issue but this code version is more stable in any case.