Row selection should be on a tap event instead of a touchstart
If the event is on a touchstart the handler is called as soon as the user touch the screen. This may be not the expected behaviour on mobile, because if the user is trying to scroll the content of the table, the event is fired to soon and this might not be the intention of the user.