DataTable: Fix rows not updated by model since v1.1
Since PrimeNG 1.1.2 (and with Angular 2.4), updating the data model of a
DataTable doesn't update the table rows that are already present. For
instance, if the data model is:

```ts
this.cars = [
    {"brand":"Brand 1","year":2001},
    {"brand":"Brand 2","year":2002},
];
```

and an event changes this data to:

```ts
this.cars = [
    {"brand":"Brand 3","year":2003},
    {"brand":"Brand 4","year":2004},
    {"brand":"Brand 5","year":2005},
];
```

the table will still display `2001` and `2002` rows, and append `2005`
at the end.

Demonstration of the bug in a Plunker:
http://plnkr.co/edit/6BvfsmCkIvXfvKYBc0hT?p=preview

Credits to Victor Ponce (@wujitouch) for the fix.

Closes #1645