dropdown: Prevent form submission on filter input
Currently when one presses the "Enter" key on the dropdown filtering
input, it submits the form surrounding the dropdown. In the following
example, the whole form is submitted (whereas the user just wants to
search in the dropdown results):

    <form (submit)="goToAnotherPage()">
	<p-dropdown [options]="cities" [(ngModel)]="selectedCity"
	            placeholder="Select a City"
                    filterPlaceholder="Search in existing cities...">
        <input type="text" [(ngModel)]="firstName" />
        <input type="text" [(ngModel)]="lastName" />
        <button>Submit</button>
    </form>

This commit prevents the form from being submitted by catching the
"Enter" key press.