Dialog fires onHide unwanted onHide event
When the dialog is created with property visible set to false, the setter for this property calls dialog#hide, which fires the event. The onHide event should only fire when a visible dialog is closed, not when it hasn't yet been visible.

The proposed change sets the property preventVisibleChangePropagation to true initially, preventing the call to dialog#hide (which fires the event) when creating the dialog. It is set back to false in dialog#AfterViewInit, which could also be done in dialog#ngOnInit if this method was added.