#3349  autoResize bug fixed
If user would like to use autoResize property, the row and column number had to be  declerated otherwise they return as undefined and autoResize had not worked.
After the fix, if user does not declare  row or column number and would liketo use autoResize, there happens a text area with default row = 5 and default column = 30 with autorResize property.