Add full accessibility support to calendar component. Demo can be found here : https://davidkirolos.github.io/primeng/#/calendar
Keyboard interaction

Left: Move focus to the previous day. Will move to the last day of the previous month, if the current day is the first day of a month.
Right: Move focus to the next day. Will move to the first day of the following month, if the current day is the last day of a month.
Up: Move focus to the same day of the previous week. Will wrap to the appropriate day in the previous month.
Down: Move focus to the same day of the following week. Will wrap to the appropriate day in the following month.
PgUp: Move focus to the same date of the previous month. If that date does not exist, focus is placed on the last day of the month.
PgDn: Move focus to the same date of the following month. If that date does not exist, focus is placed on the last day of the month.
Alt+PgUp: Move focus to the same date of the previous year. If that date does not exist (e.g leap year), focus is placed on the last day of the month.
Alt+PgDn: Move focus to the same date of the following year. If that date does not exist (e.g leap year), focus is placed on the last day of the month.
Home: Move to the first day of the month.
End: Move to the last day of the month
Tab / Shift+Tab: If the datepicker is not in inline mode, navigate between calander grid and header/grid/button bar selection buttons, otherwise move to the field following/preceding the date textbox associated with the calendar
Enter: Fill the date textbox with the selected date then close the calendar widget.
Space: When focused on prev/next month it will load the prev/next month.
Special interaction when focus on different elements

Key Down / When focus is on the input field and calendar overlay is hidden: This will open the calendar overlay.
KeyUp | KeyDown / when focus is on month/year navigator: this will load the corresponding month/year.

Fix Enter and space buttons are not working on IE and FireFox.
Add Close button to calender overlay It's required for accessibility, to enable it just set (showCloseButton=true).
Auto focus on calendar overlay now happen if:
1- User click on Show date-picker button.
2- Calendar input filed is readOnly.
3- Focus is on the input field and user hit KeyDown on the keyboard.
Add accessibility reader support for "Today","Clear","Show datepicker" and "Close" buttons.
Now "Previous Month", "Next Month", "Close" and "Show datepicker" buttons support internationalization.