Make ObjectUtils.equals() work with frozen objects
In order to avoid infinite recursion, the method added a "visited" flag
as property to the compared objects. This caused the method to raise an
error when "frozen" (read only) objects were compared that cannot be
extended. This fix uses a stack to memorize visited objects instead of
setting a property on the objects, and therefore allows frozen objects
to be compared as well.

In real life, this can happen when feeding a PrimeNG data table with
data queried using the Apollo GraphQL client (see issue #4600).

Also add unit tests to confirm proper behavior of ObjectUtils.equals().