feat(dropdown): do not close on enter if no match
Proposal:
When using the filter option, if all options are filtered, and so
no match found, I believe that keyboard enter key shouldn't close the
dropdown.
It give the wrong feeling that a new value had been selected where it's
not the case.

This is a solution to not close the dropdown in such a case.