Button Visuals Fix
In the past, a clicked button's visual "active" was overridden by "focus" , resulting a bad css. Also, on dataview's layout options, buttons became nearly invisible due to a similar problem. Now, if a button is active, it is not blocked by focus. Moreover, focused buttons are highlighted with the hue of the background colour.
This fix is focused on some free themes. Omega is not touched.