Fix for suggestions on second time load.
I'm using p-autocomplete for postcode lookup. it is work like below.
search with HU1 1UU post code it will give me 1 suggestion with placeholder 13 addresses.
When i'm selecting that I'm changing the suggetions with that 13 addresses from api result.
at that time below line is stopping me to show 13 addresses in suggestion dropdown. Specially this.loading is false so it is not entering inside of condition in handleSuggestionsChange function.

Code Line :  if (this.panelEL && this.panelEL.nativeElement && this.loading) {