Multiselect - clear input on filter cross click
We a specific requirement in the multiselect component that when we click the cross at the left hand side, rather than closing the popup, it should just clear the filter input. 

I have added a flag 'clearInputOnCross' which when will be true, will only clear the filter text input and will not close the popup.