Fixing am/pm hour validation
This fixes the issue with improper am/pm hour validation. This is shown by setting a max of say 5pm and trying to go past 5am or setting a minimum of 5 am and trying to go earlier than 5 am. Also fixes the am/pm toggle that didn't validate at all.