#7149 Clicking icon in advanced fileupload button does not open filechooser
The icon in the "Choose" button of the advanced fileupload button is
positioned absolute and is defined after the <input type="file">
element, which is also position=absolute.

So when a click is done in the region of the icon, the event is not
recorded on the file upload element, but on the icon and thus swallowed.

The solution is to raise the file element over the other elements via
a z-index style. This is considered to be save, as the button creates
a new stacking context.