#7669: SortIcon calculates aria text, but does not bind the template to it
It was noted that sortable tables do not give useful info about the
sort state for blind users. We tried to use the ariaLabel, ariaLabelAsc
and ariaLabelDesc properties, but they did not improve the situation.

Analysis in the inspector showed, that the aria-label attribute is not
updated, as I would expect it to be.

This change binds the calculated SortIcon#ariaText to the templates
aria-label attribute.