build: use ng-packagr to create APF compatible library build
For Angular 9.0.0, with ivy, projects will need to compile this
library via Angular's compatibility compiler (ngcc).

Currently the layout of the distributable for primeng is not supported
by ngcc. It is better to layout the library following the Angular
Package Format (APF) and the simplest way to achieve this is by using
ng-packakr to build the library distributable.