Added text property to allow clickable text along with icon.
Use case - As of now, we could not click on the row text shown on the cell and expand its child nodes. We can expand child nodes only on the icon click. With this change, it would be an option to show cell text as a part of anchor (clickable).