Add output to TabMenu on item selection
The tab menu currently does not let outside components know when the tab inside has changed. While it is possible to bind commands to the menu items, it could still be very useful to have the ability to react to a different tab being selected. The event emitter is named to support two way binding on the activeItem property.