Fixed #13455 - TieredMenu does not show sub menus when mouse is hovered over the parent menu item
Regression from 28989fd
To fix the problem the isDirty flag needs to be set to true when the popup is shown or an item is clicked.
This will ensure that onItemChange(..) is called when a MenuItem is hovered by the mouse, which will mark the MenuItem as active and therefore uncover the sub MenuItems (children).