properly fix issue #13361
merge pull request #13392 fixed issue #13361 but caused filters not to automatically close when clicking different filter buttons.

This change properly  fixes issue #13361 and as a result fixes the bug that was induced by the original attempted fix.

Instead of stopping the click event propagation,   the isFilterElement method is improved by calling a new  isFilterElementIconOrButton method