Fix keyboard navigation
When any tab is toggled, run fixKeyboardNavigation method to set tabIndex of focusable elements inside of tabs, both open and closed. When a tab is open, focusable elements inside of that open tab have tabIndex of 0, when it's closed, set them to -2 to avoid accessing.

Using -2 instead of -1, as DomHandler.getFocusableElements ignores elements with tabIndex of -2